+++ game/game_libs/dlls/mortar.cpp
@@ -83,7 +83,7 @@ void CFuncMortarField::KeyValue(KeyValueData* pkvd)
 	}
 	else if ( FStrEq(pkvd->szKeyName, "m_flSpread") )
 	{
-		m_flSpread = atof(pkvd->szValue);
+		m_flSpread = static_cast<float>(atof(pkvd->szValue));
 		pkvd->fHandled = TRUE;
 	}
 	else if ( FStrEq(pkvd->szKeyName, "m_fControl") )
@@ -183,10 +183,10 @@ void CFuncMortarField::FieldUse(CBaseEntity* pActivator, CBaseEntity*, USE_TYPE,
 
 		CBaseEntity* pMortar = Create("monster_mortar", tr.vecEndPos, Vector(0, 0, 0), pentOwner);
 		pMortar->pev->nextthink = gpGlobals->time + t;
-		t += RANDOM_FLOAT(0.2, 0.5);
+		t += RANDOM_FLOAT(0.2f, 0.5f);
 
 		if ( i == 0 )
-			CSoundEnt::InsertSound(bits_SOUND_DANGER, tr.vecEndPos, 400, 0.3);
+			CSoundEnt::InsertSound(bits_SOUND_DANGER, tr.vecEndPos, 400, 0.3f);
 	}
 }
